move pump setting back to plat_modbus.c
